moved simplify ids to util
